+++ apps/pipeline/lib/pipeline/reader.ex
@@ -1,5 +1,8 @@
 defmodule Pipeline.Reader do
-  @moduledoc "TODO"
+  @moduledoc """
+  Behaviour describing how to interact with system edges that produce data.
+  """
+
   @callback init(keyword()) :: :ok | {:error, term()}
   @callback terminate(keyword()) :: :ok | {:error, term()}
 end
