+++ apps/pipeline/lib/pipeline/writer/topic_writer.ex
@@ -1,10 +1,26 @@
 defmodule Pipeline.Writer.TopicWriter do
-  @moduledoc "TODO"
+  @moduledoc """
+  Implementation of `Pipeline.Writer` for Kafka topics.
+  """
 
   @behaviour Pipeline.Writer
   alias Pipeline.Writer.TopicWriter.InitTask
 
   @impl Pipeline.Writer
+  @doc """
+  Sets up infrastructure to produce messages to a Kafka topic. Includes creating
+  the topic if necessary.
+
+  Requires arugments:
+  * `instance` - Unique caller identity, e.g. application name
+  * `topic` - Topic name
+  * `producer_name` - Unique Kafka producer identity
+  * `endpoints` - Kafka broker endpoints
+
+  Optional args:
+  * `retry_count` - Times to retry topic creation. Defaults to 10.
+  * `retry_delay` - Milliseconds to initially wait before retrying. Defaults to 100
+  """
   def init(args) do
     case DynamicSupervisor.start_child(Pipeline.DynamicSupervisor, {InitTask, args}) do
       {:ok, _} -> :ok
@@ -13,6 +29,13 @@ defmodule Pipeline.Writer.TopicWriter do
   end
 
   @impl Pipeline.Writer
+  @doc """
+  Writes data to a Kafka topic.
+
+  Requires configuration:
+  * `instance` - Unique caller identity, e.g. appliation name
+  * `producer_name` - Unique Kafka producer identity
+  """
   def write(content, opts) when is_list(content) do
     instance_producer = producer(opts)
 
@@ -28,66 +51,5 @@ defmodule Pipeline.Writer.TopicWriter do
     producer_name = Keyword.fetch!(config, :producer_name)
 
     :"#{instance}-#{producer_name}"
-  end
-end
-
-defmodule Pipeline.Writer.TopicWriter.InitTask do
-  @moduledoc "TODO"
-
-  use Task, restart: :transient
-  use Retry
-
-  def start_link(args) do
-    Task.start_link(__MODULE__, :run, [args])
-  end
-
-  def run(args) do
-    config = parse_config(args)
-    producer_spec = producer(config)
-
-    Elsa.create_topic(config.endpoints, config.topic)
-    wait_for_topic!(config)
-
-    case DynamicSupervisor.start_child(Pipeline.DynamicSupervisor, producer_spec) do
-      {:ok, _pid} ->
-        :ok = Registry.put_meta(Pipeline.Registry, config.name, config.topic)
-
-      {:error, {:already_started, _pid}} ->
-        :ok = Registry.put_meta(Pipeline.Registry, config.name, config.topic)
-
-      error ->
-        raise "TODO: #{inspect(error)}"
-    end
-  end
-
-  defp parse_config(args) do
-    instance = Keyword.fetch!(args, :instance)
-    producer = Keyword.fetch!(args, :producer_name)
-
-    %{
-      instance: instance,
-      endpoints: Keyword.fetch!(args, :endpoints),
-      topic: Keyword.fetch!(args, :topic),
-      name: :"#{instance}-#{producer}",
-      retry_count: Keyword.get(args, :retry_count, 10),
-      retry_delay: Keyword.get(args, :retry_delay, 100)
-    }
-  end
-
-  defp wait_for_topic!(config) do
-    retry with: config.retry_delay |> exponential_backoff() |> Stream.take(config.retry_count), atoms: [false] do
-      Elsa.topic?(config.endpoints, config.topic)
-    after
-      true -> config.topic
-    else
-      _ -> raise "Timed out waiting for #{config.topic} to be available"
-    end
-  end
-
-  defp producer(config) do
-    {
-      Elsa.Supervisor,
-      [endpoints: config.endpoints, connection: config.name, producer: [topic: config.topic]]
-    }
   end
 end
